Factory method to create a brand new Report

+++ common/src/main/java/app/web/report/dto/ReportSummary.java
@@ -6,7 +6,7 @@ public interface ReportSummary {
     float getTotalMaterialWeight();
     float getWeightDifference();
     float getCorrectProductPercent();
-    float getIncorrectProductPcs();
+    long getIncorrectProductPcs();
     long getOverFilledProductPcs();
     long getNotFulFilledProductPcs();
 }
